requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.4b2
click==7.1.2
coloredlogs==15.0
Django==3.2
django-extensions==3.1.3
django-filter==2.4.0
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Faker==8.1.2
greenlet==1.0.0
humanfriendly==9.1
Markdown==3.3.4
msgpack==1.0.2
mypy-extensions==0.4.3
neovim==0.3.1
pathspec==0.8.1
phonenumbers==8.12.22
Pillow==8.2.0
PyJWT==1.7.1
pynvim==0.4.3
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
regex==2021.4.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
